# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SocialFlow AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces:
    get:
      operationId: listWorkspaces
      tags: [workspaces]
      summary: List user workspaces
      responses:
        "200":
          description: List of workspaces
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workspace"
    post:
      operationId: createWorkspace
      tags: [workspaces]
      summary: Create workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
  /workspaces/{id}:
    get:
      operationId: getWorkspace
      tags: [workspaces]
      summary: Get workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "404":
          description: Not found
    patch:
      operationId: updateWorkspace
      tags: [workspaces]
      summary: Update workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
    delete:
      operationId: deleteWorkspace
      tags: [workspaces]
      summary: Delete workspace
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── SOCIAL ACCOUNTS ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Workspace:
      type: object
      required: [id, name, ownerId, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        ownerId:
          type: integer
        createdAt:
          type: string
          format: date-time
    WorkspaceInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        logoUrl:
          type: string
    WorkspaceUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        logoUrl:
          type: string

    # ─── SOCIAL ACCOUNT SCHEMAS ────────────────────────────────────────────────
...